Server Backend/requirements.txt
Flask==0.12.1
azure-common==1.0.0
requests==2.14.2
opencv-python==3.2.0.7
Flask-Cors==3.0.2
--find-links wheelhouse

